Isolate libreoffice invocations from each other
This addresses AE-1053

It seems to suffice that an unique UserInstallation parameter is used
for each invocation. The parameter does not actually have much to do
with the installation location, but it tells where the settings of the
user are located. Names of IPC sockets are derived from this path.