Survive absence of LibreOffice configuration
Normally, if the LibreOffice configuration directory is present, it
will be used to initialize the temporary UserInstallation
(i.e. config) directory for a single LibreOffice run.

Previously the code would fail with an IOException if the
configuration directory is absent. With this change, the absence will
be silently ignored. This is mainly to make it more convenient to test
the backend.