AE-1406 Always handle :include-kasitelty
The previous ketju search queries had a trailing OR branch that would
ignore the parameter, making it impossible to hide processed ketjus in
some combinations of parameters.

The queries are rephrased so that it is hopefully easy to see that the
three available constraints can be applied independently from each
other.