AE-1577 Sum of two heat losses instead of only one
The 2018 ET has two values for describing heat losses related the heat
distribution system:
- [:lahtotiedot :lammitys :tilat-ja-iv :lampohavio-lammittamaton-tila]
- [:lahtotiedot :lammitys :lammin-kayttovesi :lampohavio-lammittamaton-tila]

Before this change, only the first one of these was placed in the A9
cell of the 8th page of the ET PDF. This change adjusts it so that
their sum is used instead. A10 still gets the latter value.