AE-1630 Add missing check for deleted valvonta
Paging in käytön valvonta was falling over a bit, because the count of
results was not counting with exactly the same search conditions as
the query that would fetch the results.

The code layout was also changed so that the where keyword is alone on
its own line and all the query conditions are in a separate indented
section.