AE-1638 Extra searches for oikeellisuuden valvonta
The keyword search and search by type of last toimenpide are similar
to what were implemented for Käytönvalvonta in AE-1637. In addition,
there is a search by laatija-id of the Energiatodistus under valvonta,
and search by käyttötarkoitusluokka of the ET. KTL search has a logic
similar to what is used in the public statistics retrieval.