AE-1738 Close cases locally even if Asha gives an error
A known reason for error when asking Asha to close a case is that the
case has already been closed in Asha. This has prevented pääkäyttäjä
from closing cases in ETP. This change makes the close command
progress to completion even if Asha gives an error to the request to
close the case.