AE-1794 Add POST route to close käytönvalvonta without Asha
To deal with cases where states of ETP and Asha diverge, this change
adds an API that processes a case-closing toimenpide without
contacting Asha at all. This can be necessary e.g. if pääkäyttäjä
manually closes the case in Asha for some reason.

The API is kept similar to the usual one used POSTing toimenpiteitä,
but there is a check that only permits toimenpide of type :closed.