Add Swedish translations to Suomi.fi messages regarding toimenpide
Added Swedish translations to Suomi.fi messages that are sent when a new
toimenpide is added to valvonta.

Also fixes a problem with Finnish messages. Previously the postal
district was not filled in the message.

The title of the message is in Finnish, and Swedish message is after the
Finnish one, as it is in the email.

Implements: AE-1816