Update WSS4J
WSS4J is the library used for signing soap messages sent to Suomi.fi.
The previous implementation was apparently made using the official
example, that to this day still uses the obsolete version of WSS4J.

The details of what should happen in signing are scarce. Created a test
that compares a document signed with the previous version of the library
to a new one. Cryptographic elements cannot be compared easily, so the
test instead just checks that all the correct elements are in place.