Upgrade Flyway
In 9.1.4 Flyway changed to use transactional locks, but it causes
migrations to freeze. Reverted to old default.

In 9.19.0 breaks things. Audit tables create a log of added kayttaja
entries. The audit log try to use the function current_kayttaja_id(),
which relies on "application name" being set. It's set in
beforeMigration.sql, but in this version callbacks have a separate
connection. Added ApplicationName to connection url to get around the
problem.

Implements: AE-1909