Move palveluväylä API to routes root and version it
Because of technical reasons in the sidecar (liityntäpalvelin), we need
have the palveluväylä API start without dashes.  Solved the issue by
moving the API to server root.

(Technical reason: Our documentation is the autogenerated OpenAPI
specification. The relative URLs there are from the server root,
previously beginning with /api/palveluväylä/. In the sidecar
configuration, our service is defined as basepath service, with the url
http://instances.core-public.local:8080/api/palveluvayla. It's defined
that way, so that the users cannot access APIs other than those under
/api/palveluvayla. However the url that users use, is in the form of
https://sidecar/r1/organisation/identifier/codes/service-code/url-from-openapi-specification.
The service code is defined on the sidecar, and cannot contain slashes.
If we use palveluvayla as the service code, and as the root of our
palveluväylä API, the urls that are used to connect our service using
palveluväylä match those that are defined in the OpenAPI specification.)

Version the palveluväylä API so we can make breaking changes in the
future.