AE-2042 Special handling to exec LibreOffice on Windows
One difference on Windows is that the executable program is not
made available with the usual libreoffice name. This change makes
the backend just call it directly from its most common location.

Naive attempts at using the UserInstallation override resulted
in errors about a corrupted bootstrap.ini file. We do not absolutely
need to run multiple PDF generations in parallel, so the issue
is sidestepped by letting go of the mechanism needed for multiple
parallel executions.